Add .toSynchronous() to slime.old.Loader
Also:
* Massive refactoring to better separate old loader code
