+++ loader/jrunscript/expression.js
@@ -915,6 +915,7 @@
 					this.factory = void(0);
 					this.Child = void(0);
 					this.get = void(0);
+					this.toSynchronous = void(0);
 
 					var source = adaptLoaderArgument(p);
